style dogList and dogCard with tailwind
