+++ src/components/DogsList.tsx
@@ -1,17 +1,15 @@
 import React from 'react';
 import dogs from "../data/temp-data.json"
-import Dog from './Dog';
+import Dog from './Dog/DogCard';
 
 function DogsList() {
 
   return (
-    <div className="DogsList">
-        <ul>
-          { dogs.map((dog) => {
-            return <Dog dog={{...dog}}/>
-          })}   
-        </ul>
-    </div>
+    <ul className="w-full flex container">
+      { dogs.map((dog) => {
+        return <Dog dog={{...dog}} key={dog.id}/>
+      })}   
+    </ul>
   );
 }
