+++ src/data/temp-data.json
@@ -7,6 +7,7 @@
             "RED",
             "WHITE"
         ],
+        "breed": "pitbull mix",
         "weight": 60,
         "birthdate": "2018-3-14T18:30:00"
     },
@@ -18,6 +19,7 @@
             "TAN",
             "WHITE"
         ],
+        "breed": "pitbull mix",
         "weight": 40,
         "birthdate": "2020-12-21T18:30:00"
     }
